layout/forms/test/test_button_inner_events.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=843003
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 843003</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 843003 **/

  SimpleTest.waitForExplicitFinish();

  var currentEvent = "";
  var buttonEvents = 0;
  var spanEvents = 0;
  var b = null;
  var s = null;

  function buttonEventHandler(e) {
    is(e.type, currentEvent, "should get the expected event");
    buttonEvents++;
  }

  function spanEventHandler(e) {
    is(e.type, currentEvent, "should get the expected event");
    spanEvents++;
  }

  function setup(name) {
    currentEvent = name;
    b.addEventListener(currentEvent, buttonEventHandler);
    s.addEventListener(currentEvent, spanEventHandler);
  }

  function test(expectedButtonEvents, expectedSpanEvents) {
    is(buttonEvents, expectedButtonEvents,
       "button should have received " + expectedButtonEvents + " " +
       currentEvent + " events");
    is(spanEvents, expectedSpanEvents,
       "span should have received " + expectedSpanEvents + " " +
       currentEvent + " events");
  }

  function clear() {
    b.removeEventListener(currentEvent, buttonEventHandler);
    s.removeEventListener(currentEvent, spanEventHandler);
    currentEvent = "";
    buttonEvents = 0;
    spanEvents = 0;
  }

  SimpleTest.waitForFocus(function() {
    b = document.getElementsByTagName('button')[0];
    s = document.getElementsByTagName('span')[0];

    setup('focus');
    b.focus();
    test(1, 0);
    clear();

    setup('blur');
    b.blur();
    test(1, 0);
    clear();

    setup('focus');
    s.focus();
    test(0, 0);
    clear();
    is(document.activeElement, document.body, "focus is back to body");

    if (!navigator.platform.startsWith("Mac")) {
      setup('focus');
      synthesizeMouseAtCenter(b, {});
      test(1, 0);
      clear();
    }

    setup('mousedown');
    synthesizeMouseAtCenter(b, {});
    test(1, 1);
    clear();

    setup('mouseup');
    synthesizeMouseAtCenter(b, {});
    test(1, 1);
    clear();

    setup('click');
    synthesizeMouseAtCenter(b, {});
    test(1, 1);
    clear();

    b.focus();

    setup('keydown');
    synthesizeKey("VK_A", {});
    test(1, 0);
    clear();

    setup('keyup');
    synthesizeKey("VK_A", {});
    test(1, 0);
    clear();

    setup('keypress');
    synthesizeKey("VK_A", {});
    test(1, 0);
    clear();

    setup('DOMActivate');
    synthesizeMouseAtCenter(b, {});
    test(1, 0);
    clear();

    synthesizeMouseAtCenter(b, { type: 'mouseover' });
    is(b.mozMatchesSelector(":hover"), true, ":hover should match");
    is(s.mozMatchesSelector(":hover"), true, ":hover should match");

    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=843003">Mozilla Bug 843003</a>
<p id="display"></p>
<div id="content">
  <button>
    <span>foo</button>
  </button>
</div>
<pre id="test">
</pre>
</body>
</html>
